backend/pyproject.toml
[project]
name = "book-rag"
version = "0.1.0"
description = "RAG System for Book Content Question Answering"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "qdrant-client>=1.7.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.5.0",
    "langchain>=0.0.300",
    "langchain-community>=0.0.20",
    "python-multipart>=0.0.6",
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
